Introduce concept of a SpanInitiator
In order to allow for more powerful integrations with this library,
the notion of a SpanInitiator is being extracted from the SpanBuilder.
The idea behind this is that it provides an intercept point for
external logic to be integrated into the span creation process.

For example, the existing SampleController concept results in
an UnsampledRootSpan when there is no active span and a span
is being created which is determined to not be sampled.  This behavior
can be entirely encapsulated by a SpanInitiator implementation
and externalized from this library altogether.

Similarly, the ability to gain knowledge of all spans being created
and or potentially decorate these span instances opens up a lot of
possibilities for consumers of the API.

As a result of the above, the SampleController support will be removed
in a subsequent commit.